[aes, rtl] Clear register shares with independent PRNG bits
Previously, the bits used for clearing the second share of registers
were obtained by applying a linear permutation to the bits used for the
first share. Depending on the used permutation, this could lead to some
bits in both shares always being cleared to the same value, which could
result in SCA leakage. This commit thus changes the design to split the
LFSR output and use seperate bits for both shares.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>