[tlul/rtl] Increase max devices per `tlul_socket_1n` to 64
This is necessary for attaching up to 10 mailboxes to Ibex and the SBA,
in addition to the many other existing devices.

9e6ceac884189ffd38b4665a888a04218d67368b increased this from 16 to 32
without accompanying changes, so no further action seems to be required.
It should be noted that such wide 1-to-N sockets may result in overly
long combinatorial paths -- but we'll have to see timing results to know
whether this is an actual problem.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>