[top_darjeeling/soc_proxy] Reorganize pinmux config
This updates the pinmux configuration and makes most IOs
dedicated.

Unfortunately the pinmux matrix must have at least 1 input and one
output connection to go through it with the current tooling.

To that end, the number of SoC GPIOs is increased to 16 so that 4
of them (for I/O each) can be routed through the remaining muxing
matrix.

Co-authored-by: Andreas Kurth <adk@lowrisc.org>
Signed-off-by: Michael Schaffner <msf@opentitan.org>