[hw,dma,rtl] Only capture return data in read state
rsp_valid is used to signal a valid response and acknowledges a write.
Previously, this signal led to capture the return data in both cases.
However, during a write operation, this leads the DMA to capture arbitrary
bus data, which was active during a write operation. This wrong data was
eventually fed into the SHA engine, leading to a wrong hash.

In this PR, we guard the capture_return_data signal, with the read
states of the FSM, to only capture data there.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>