[keymgr_dpe, lint] Fix lint for `keymgr_dpe_ctrl` module
This commit removes the waivers that were temporarily
added during RTL development of keymgr_dpe.

The following RTL fixes are made to fix lint errors that
appear after removing the waivers:

* Remove `dis_req`, as it is not used. Previous keymgr
defines a similar variable `dis_op`, which is always asserted
when an operation that is not {advance, generate, id-generate}
is selected. This counter-measure makes sure that a random
op_start_i moves FSM to disabled. Keymgr_DPE does not
implement such counter-measure yet (and it's not clear if it
should), so this signal is removed.
* Simplify the type of `destination_slot` to only validity bit
* Mark `exportable` policy bit unused.
* Let `StCtrlDpeWipe` state keep active operations, so that
the KMAC communication is not hung problematically.

Signed-off-by: Fatih Balli <fatihballi@google.com>