WIP [aes] Re-enable switching off the masking at run time on the CW305
On the CW305, the entropy complex isn't instantiated due to resource
constraints. As a result, the magic seed required by the PRNG to produce
the all-0 output cannot be loaded into AES.

This commit reverts the SecAllowForcingMasks parameter to directly
drive the output of the PRNG to 0 in case  the FORCE_MASKS bit in the
auxiliary control register is set to 1.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>