[keymgr_dpe, dv] enables keymgr_dpe exp keyslot checks
In the keymgr_dpe_if adds an internal_key_slots array that's
assigned in the tb to the dut.u_ctl.key_slots_q signal. This
allows the scb and intf to check the expected key slot values
to the actual key slot values for a given slot index.

In the scb the process_kmac_data_rsp() function will update the
expected key_slot values for the dst slot upon updating the internal
key. In the intf a function is added compare_internal_key_slot()
that compares the provided dst key slot to the dut's.

Signed-off-by: Bobby Reyes <bobby@rivosinc.com>