[util] Get rid of "type" field in Modes
The existing python code was using a string to tell the base class how
to look up sub-modes for an instance of a given derived class. This
was a bit complicated!

There are actually only two places where this does anything anyway, so
duplicate them and get rid of the string magic.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>