[keymgr_dpe/rtl] Take max key version into account for root key
Technically it's possible to generate output keys from the root key
(i.e., the key with boot stage 0), so it is meaningful to allow SW to
set the maximum key version when invoking the advance operation that
loads the root key.  Prior to this commit, `keymgr_dpe` would ignore the
`MAX_KEY_VER_SHADOWED` CSR for the root key and always set its maximum
key version to zero -- that's now fixed.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>