[dma/dv] Fix computation of integrity bits
Prior to this commit, only the `rsp_intg` field (i.e., the integrity bits
associated with the response opcode, size, and error) was computed, and
that happened in the `update_mem` function.  The problem is that the
`data_intg` field (i.e., the integrity bits for data) also needs to be
computed when randomizing the response (i.e., in the `randomize_rsp`
function), and that `update_mem` modifies the data, so its integrity
bits need to be recomputed.

This commit fixes those problems by computing all integrity bits when
the response is randomized (in `randomize_rsp`) and by recomputing the
data integrity bits when the data gets changed in `update_mem`.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>