[keymgr,keymgr_dpe] Change reseed counter to count PRNG updates instead of cycles
Previously, the reseed counter was counting clock cycles which was
problematic for the following reasons:
- Once started, the counter couldn't be stopped anymore and keymgr would
  repeatedly request entropy even if being completely idle. This is bad
  for power.
- The number of clock cycles since the last reseed operation is no good
  indication for the seed lifetime.

Therefore, this commit changes the design to only advance the counter
when the PRNG is updated. Such updates only happen while keymgr is busy
and the number of PRNG updates is a more accurate measure for the seed
lifetime.

Co-authored-by: Andreas Kurth <adk@lowrisc.org>
Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>