[chip-test,sival] Optimize ENTROPY_SRC config for ast_clk_rst_inputs
This test previously configured the ENTROPY_SRC in Firmware Override:
Extract and Insert mode and erroneously enabled the routing of
entropy to firmware (ENTROPY_DATA register). Also the entropy source
was reconfigured multiple times during the test.

This commit modifies the test to 1) use the Firmware Override: Observe
mode in which entropy is observable from the Observe FIFO but
also continues to flow through the pipeline to CSRNG and the EDNs, and
2) to not re-configure the entropy source but just restart it to restart
the entropy re-collection.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>