[dv] Chop reset delay by a factor of 100 in cip_base_vseq
This is a bound on the time we'll wait before we try to inject a reset
in the middle of a stress sequence. 10 million cycles is... quite a
long time!

This cuts it down to 100k cycles, which is still surely long enough to
get to an interesting point in the state space of the device.

But rather more quickly.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>