[jtag,dv] Correct FSM driving after trst_n
The previous version of this code was assuming that we'd come out of
trst_n in FSM state RunTest/IDLE. This *was* the behaviour of the
debug module with a previous version of the RTL (which is presumably
how this happened). But it's not what is supposed to happen, and it's
not what the RTL does any more (since commit 663be282861).

Explicitly describe what we think should be going on and control tms
to be zero across the clock edge, which ensures we end up in
Run-Test/Idle regardless of where we started.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>