[sw,dv] Add basic `status_t` (%r) logging in DV
Adds a simple substitution rule to the current C->SV logging backdoor.
This rule replaces the "%r" and "%!r" custom formatting specifiers (used
to print a `status_t` normally and in a JSON representation) with a hex
dump of the 32-bit packed value. This prevents errors in DV tests when
using `status_t` logs.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>