[entropy_src/dv] Predict entropy drops in the scoreboard
This commit adds prediction for entropy drops to the scoreboard.
These drops can happen in front of the postht FIFO or in front of
the esbit FIFO. To this end the scoreboard now contains a task
for each FIFO to be able to predict the backpressure (the final FIFO
and the sw read FIFO are not included).

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>