[rv_dm,dv] Avoid injecting extra resets in rv_dm_tap_fsm_vseq
Doing so causes problems in the stress_all_with_rand_reset test. When
run_seq_with_rand_reset_vseq spawns the vseq to run until the reset,
it sets do_apply_reset = 0 to avoid the vseq injecting a reset itself.
We need to pass that along here.

I'm not completely convinced by this vseq in general. It was added in
3131e9aabc5 to improve TAP FSM coverage. I strongly suspect we can
ditch it, but that's probably an issue for a future commit.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>