[dv] Allow assertions from ASSERT_FPV_LINEAR_FSM to complete
The assertion is supposed to start and say that we don't come back to
a state until reset. Unfortunately, the implementation was disabling
the property when reset happened, which meant that the assertion never
finished, so it got marked as incomplete in the coverage.

Change the call to `ASSERT() so that it doesn't kill the property any
more, relying on the fact that the property will terminate on its own
once we go into reset.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>