[ci] Check that the FPGA jobs do not run duplicates and are exhaustive
With the filter tags filtering logic, it is very easily to create a
situation where some tests are run in several jobs, or to forget running
tests because of a non-exhaustive set of filters.

This commit adds a final CI job that checks the above condition. This
is achieved by downaloding the list of tests run by every job and then
making sure that there are no duplicates. We also check for missing tests
by running a query on all non-broken, non-manual, non-skip_in_ci FPGA tests.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>