[cert] move cert keygen function to otbn_boot_services lib
This moves the `cert_ecc_p256_keygen()` function from the `cert` lib to
the `otbn_boot_services` lib to simplify on-host unit testing of the
`cert` lib. The `otbn_boot_services` lib is only tested on-device at the
moment due to the dependencies on specific OpenTitan hardware that is
difficult to mock. In refactoring this, the cert_unittest is now fixed.

Signed-off-by: Tim Trippel <ttrippel@google.com>