[mem_bkdr_util] Support varying No. of PRINCE half rounds for scrambling
Previously, all scrambled memory primitives had to use the same number
of PRINCE half rounds which was set to 3. This commit adds a new and
optional argument to the mem_bkdr_util constructor, to allow configuring
different values. If the argument is not provided, the default of 3
PRINCE half rounds is used.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>