[prim,fpv] Get rid of ASSERT_KNOWN in an FPV context
Unknown signals aren't really supported by Jasper, which drops the
assertion accordingly. Unfortunately, this shows up as a coverage
hole (for a rather silly reason). Get rid of the assertion entirely if
we are in this context (the signal is not X because the tool doesn't
allow that to happen...)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>