[prim,fpv] Remove initialisation before a test
This initialisation isn't really needed because the array will come up
as 'x and the SboxPermutationCheck_A assertion below would fail if the
end result wasn't '1.

The existing code also causes a reported coverage hole in the FPV
flow, which notes that assigning sbox_perm_test to zero has no effect.

Obviously, this is a little silly: it's inside ifndef SYNTHESIS so
coverage doesn't really matter! But waiving this sort of coverage hole
more generally might be tricky and this particular case can just be
"fixed".

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>