[prim,lint] Remove waiver for prim_sram_arbiter.sv
This came from b4919e58, the initial commit to prim.waiver in 2020. As
far as I can tell, it didn't even apply to the prim_sram_arbiter.sv in
the initial commit in the repository (there was only one rsp_error*
signal, but this was an output port which just matched the sram_rerror
input port).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>