[ci, sival] Make testplan validator need bazel if si_stage != NA
Modifies the schema used for testplan validation to add a requirement
for the bazel property to be defined for a given testpoint if its
`si_stage` attribute is set to any value other than `NA`.

Note that this is a one-way implication; a testpoint can have a `bazel`
property with a non-`NA` `si_stage`, as is used in a couple of cases for
e.g. manufacturing tests.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>