[rv_dm,dv] Add a task that injects TL errors to the smoke sequence
This is to fill a coverage hole, where the rom_intg_error signal is
never seen going high in the coverage.

The signal *does* go high in the rv_dm_tl_intg_err test, but that test
uses the "cover_reg_top" build_mode (in tl_access_tests.hjson). This
build mode is intentionally configured not to collect coverage for
anything but the TL-focused modules. This is done for VCS in
cover_reg_top.cfg. The point is that if the block doesn't have a model
against which we're checking, then the register tests might trigger
weird behaviour that they don't notice. Therefore we shouldn't be
tracking coverage of the non-TL-related signals for these tests,
because they might not be testing the blocks that they exercise.

In the very specific case of TL integrity errors, it doesn't really
matter: the only behaviour we expect is a fatal alert. This *is*
tested by the register tests! But making it happen very occasionally
in the smoke sequence is probably easier than reasoning about what
coverage should be included in the existing generic tests. Do that.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>