[opentitantool] improve user experience
When 'opentitiantool image manifest update ...' fails due to any input
or output file name mistyped, or any of the input files does not
exist, the error message is the same:

  Error: No such file or directory (os error 2)

Which does not help much the operator. This patch adds some code which
will print the name of the file causing the error.

Tested by invoking 'image manifest update' with incorrectly specified
various input files.

The same improvement is being made to the bootstrap option, file name
is reported if file read attempt failed.

Signed-off-by: Vadim Bendebury <vbendeb@google.com>