[top_englishbreakfast] Add second UART
For penetration testing, we've been using a second UART on the CW310 and
the silicon platform such that the host can directly communicate with
the target using the uJSON protocol instead of looping this
communication through the simpleserial channel with the scope
(available on Husky only anyways).

This commit now adds a second UART to the English Breakfast top level
such that we can re-align the capture setups again. Ultimately, this
will allow removing the simpleserial-based communication which will
quite notably simplify code maintenance in the future.

For reference, see also lowRISC/ot-sca#235.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>