[reggen] Move Register.make_multi into MultiRegister
The "multi-register stuff" now lives in MultiRegister.from_raw and the
bit that has more to do with collecting up registers has moved into
the new Register.collect_registers function.

This is supposed to support a rather more general situation, where
we're pasting together several registers. Honestly, it's a bit too
general! This explains a bunch of assertions that check the pasted
registers are compatible. Those could probably be weakened with a bit
of thought, but the call site doesn't need that yet.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>