[spi_host, dv] : Add missing argument for format specification
This commit fixes a build issue seen when running the block using the
VCS simulator. In one of the sformatf statements in the
spi_host_base_vseq, the format string contains a format character but
missed specifying the corresponding argument. This commit fixes the
issue by adding the appropriate argument corresponding to the format
character.

Signed-off-by: Venkat Krishnan <venkateswarank.opentitan@gmail.com>