[rom_ctrl, dv] Fixed rom_ctrl checker fsm conditional coverage
The conditional coverage was uncovered on L157 for kmax_err_i being 1.
This is due to the scenario an unaligned transitions b/w fsm state
(ReadingHigh), kmac_done_i, counter_done and kmac_err_i. All should be
synced and raised on the exact clock edge. kmac_err_i was unsynced and
raised to 1 when the state was RomAhead.

The fix tweaks the rsp_delay in rom_ctrl_env_cfg.sv just to get the
precision in the delay and that kmac_err_i should be raised for the
particular scenario when state is ReadingHigh and both kmac_done_i and
counter_done are set.

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>