[sival, keymgr] State-agnostic chip_sw_key_derivation_test
keymgr_key_derivation_test_fpga_cw310_sival_rom_ext test was marked as
broken in #21706. This commit fixes this and addresses #21500.

The source of the bug is due to the nature of the ROM_EXT that runs
through a CDI attestation flow advancing the key manager state to
"OwnerRootKey" in the process. This was not accounted for in the test,
which assumed an active "CreatorRootKey" state (this works for sim_dv
tests but not with ROM_EXT). The solution to this issue lies in
rendering the test state-agnostic, meaning it will execute its
verifications independently of the current key manager state. This
approach thus mimics the keymgr_sideload_{aes,kmac,otbn} tests which
follow a similar strategy.

Signed-off-by: Andrea Caforio <andrea.caforio@lowrisc.org>