[aes] Add GHASH module for AES-GCM
This commit adds a first implementation of the GHASH module required for
AES-GCM support. This first version uses a single, pipelined GF(2^128)
multiplier and 3 128-bit registers for the GHASH state, the hash subkey
and the encrypted initial counter block J_0 (= S). The latency of the
GF multiplier is matched to the latency of the cipher core.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>