[manuf] Finalize {Creator,Owner}SwCfg at end of perso flow
This modifies the perso flow to complete the provisioning of the
CreatorSwCfg and OwnerSwCfg OTP partitions at the end of the process.
This ensures that the perso firmware can be reentrant in case of
unexpected resets during personalization.

Previously, these partitions were provisioned earlier in the flow.
However, since we are going to bind the perso firmware with
`CREATOR_SW_CFG_MANUF_STATE`, we need to ensure that this field is
provisioned as late as possible.

Since CreatorSwCfg and OwnerSwCfg partitions are now not fully
provisioned in the OTP during the UDS certificate generation process,
overwrite the fields that are not provisioned until the end of
personalization process with the values from the OTP image. At the end
of perso flow, compare the OTP measurement used during certificate
generation with the digest stored in the OTP. Ensure that the UDS
certificate was generated using the correct OTP values.

Signed-off-by: Anthony Chen <antchen@google.com>