[aes] Align interface of clearing PRNG with masking PRNG
For historical reasons, the clearing PRNG had a req/ack interface for
both the pseudo-random data and the reseed interface. The former is
actually not required as the PRNG can always provide suitable pseudo-
random data, even in case of an outstanding reseed request. This commit
thus simplifies the req/ack interface to a single update signal, similar
to what is used for the masking PRNG.

In addition to simplifying the code, this enables re-using the main
control FSM state previously used for performing the handshake with the
clearing PRNG, to perform an actually required handshake with another
functional block such as the GHASH block used for AES-GCM.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>