[reggen] Add unpacked_dimensions field to parameters
This allows you to specify unpackage dimensions which are placed after
the name of the parameter, e.g.

parameter type_t param_name [2][3]

[2][3] would be the unpacked dimensions.

This is required to support the new method of specifying PMP reset
values via paramters for Ibex.

This commit just adds the field to the BaseParam class and allows reggen
to read it, it doesn't use the new field for anything.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>