[dv,alert_esc_agent] Wait on both clocks in wait_ping_timeout
This wait is supposed to be waiting for a while and then concluding
that the receiver has timed out (oh dear!). But the receiver might be
running on a slower async clock, so we shouldn't be timing it based on
a faster one. Make sure to wait for the slower clock.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>