[aes/rtl] Optimize GHASH state restoring
Previously, we restored a previously saved GHASH state by first
overwriting the initialized state with the masked cipher core input and
second subtracting Share 1 of S to the GHASH state afterwards.

This commit changes the design to overwrite only Share 0 of the
initialized state with the unmasked saved state and leave Share 1
untouched. As Share 1 of the state is already initialized to Share 1 of
S, no further operations have to be done. This allows saving one 128-bit
multiplexer input (or 300 GE).

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>