[sival,i2c] Fix and re-enable I2C target test
The I2C target test was failing because the ujson I2cTestConfig command
added for clock stretching did not have a CRC added to it, causing the
deserialization to fail. This commit adds this CRC, adds an additional
CRC to the configuration payload itself, and marks the test as no longer
being broken.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>