[aes/rtl] Optimize GHASH state restoring, align masked and unmasked impl
This commit aligns the unmasked and masked implementation of the GHASH
block regarding state saving and restoring. In both cases, the sum of
the GHASH state and S are saved. After restoring, S is again subtracted
also for the unmasked implementation. This allows saving one 128-bit
multiplexer input and it allows exchanging saved states between
unmasked and masked implementations which simplifies DV.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>