[sival, keymgr] Fix keymgr sealing/attestation test in unmasked envs
The `chip_sw_keymgr_derive_{sealing,attestation}` test would fail on
FPGAs with masking disabled due to faulty logic that checks whether
derived keymgr outputs differ between states by not taking into
account that the second output share is always 0 when masking is
disabled.

This commit resolves this issue by flagging two outputs as different
if and only if they differ in at least one share instead of requiring
both shares to be different which was the cause of the failure.

Signed-off-by: Andrea Caforio <andrea.caforio@lowrisc.org>