[ci] Create GitHub groups in log outputs for FPGA tests
The log are quite long and GitHub provides a feature where the
output group can split lines into groups. This is achieved by
using the --run_under command of bazel test. This allows to
run a script that then invokes the test runner. This only
thing this script does is print the grouping commands.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>