[dif] Fix dif_<ip>_irq_get_type to handle bad arguments
These functions perform an incorrect check on the irq number.
In particular, this number could be negative or above max_irq+1.
If this happens, the function will perform an out-of-bound array
access.

One particular instance where this could happen is in the isr
testutils where it can be called with an incorrect argument.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>