[bazel] Add new macros to describe hardware IP and tops
Those macros record the IPs/top description in a struct. This
struct can then be used to implement various build system
rules and macros that need access to the top description.

The main reason why these are macros and not rules is that some
of this information needs to be available during bazel's loading
phase, where only macro and variable expansion are available.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>