[make_new_dif] Only consider earlgrey IPs for isr_testutils
We are current in a weird spot: we have DIFs for IPs in Darjeeling
but isr_testutils it not quite ready to be multitop. This is because
it contains code for *every IP* but clearly not all IPs exist in
every top. This kind of works because we generate C headers for all
IPs independently of the top and so we happily compile DMA code on
Earlgrey...
However, when we start marking headers as only compatible with the
active top, this will break down and isr_testutils will not compile
anymore.

Furthermore, the isr_testutils template already assume Earlgrey by
including `top_earlgrey.h` and the Python generator does not even
take the top description as input so it's top independent.

This will require a proper fix eventually but this commit
implements a stopgap: it only considers Earlgrey IPs for inclusion
in isr_testutils. This at least guarantees that it create a
consistent, if earlgrey-specific, file.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>