[bazel] Patch rules_foreign_cc to clean build logs
Move rules_foreign_cc from the root MODULE.bazel to its own
`third_party/foreign_cc` directory so that we can patch it to fix an
issue that has not yet been supported upstream yet, in the same way as
we do for e.g. `rules_rust`.

When `rules_foreign_cc` builds using tools such as e.g. Configure+Make,
output files Configure.log and BootstrapGNUMake.log are generated and
emitted which contain references to the absolute path. When building
inside Bazel's sandbox, this produces a non-determinstic output which
changes between each run, as a result, rebuilding OpenOCD will then
change the inputs of any Bazel actions that use OpenOCD, resulting in
a reduced number of cache hits and increasing build times. By making
this determinstic, we improve reproducibility and potentially reduce
build times.

To fix this, this PR introduces a simple patch to completely remove
all the content of this logging output, as it is not needed by any later
steps as part of the build.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>