[aes,sca] Add AES-GCM FvsR batch test
This commit adds the AES-GCM fixed vs random batch test.
The host configures:
- Number of batch operation s(max 256)
- Where to trigger
- The fixed IV & fixed key
- The number of AAD and PTX blocks
- The fixed AAD and PTX blocks

The FvsR set consists of the IV and the key, i.e.,
on the device a fixed and a random set of these variables
are created.

The AES-GCM operation is executed batch_operations. The XOR
of all tags / ctxs are transmitted back to the host for
verification.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>