[aes_gcm,sca] Adapt sleep cycles
This commit adapts the number of cycles we are putting Ibex into
sleep. The previous cycle count was too large as for the mcycle
measurement the AES_TESTUTILS_WAIT_FOR_STATUS function was used
to wait for the IDLE bit. By directly using the DIF to wait for
this bit, we are reducing the overhead caused by this library.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>