Revert "[opentitantool] Refuse downgrading HyperDebug firmware"
This reverts commit 7a0134a9b4d5620435b9750683083662acc7d6ab, such that
opentitantool no longer requires the `--force` switch in order to
downgrade firmware with `opentitantool transport update-firmware`.
Instead, opentitantool will be back to the original behavior of
overwriting firmware in any case when the version number does not match
exactly.

The original reason for not downgrading was that our branches had
diverged, such that they had different firmware versions, resulting in
CI machines constantly upgrading/downgrading as CI runs from different
branches were scheduled.

PR #25917, however introduced a faulty HyperDebug firmware with never
version number, which is now on a number of CI machines, and will not be
downgraded automatically, with the result that "innocent" other PRs will
have CI failures.

I think it is clear that we should avoid such excessive flashing by
keeping the HyperDebug firmware version in sync between branches.  (We
do not need to keep all of opentitantool logic in sync, just the
built-in firmware version.)  For now, in order to avoid CI failures, we
should merge this PR, and accept some upgrading/downgrading.

Change-Id: I57739777ab3d86fa7e0375834cf9140e8e6ac984
Signed-off-by: Jes B. Klinke <jbk@chromium.org>