[tlul,rtl] Merge u_sramreqfifo and u_sramreqaddrfifo
Having two FIFOs with pointers that are kept in lock-step is a little
silly. Turn them into a single wider FIFO.

Note that the two FIFOs had compatible parameters. The address FIFO
had OutputZeroIfEmpty set to 1, which is the default value (so
u_sramreqfifo got it too). Request it explicitly in case the default
changes in future.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>