[opentitantool] Erase header of other slot during rescue
I believe I have seen cases where the slot not being flashed during
rescue still booted afterwards, meaning that the previous code flashing
and empty image to the other slot may be ineffective.  This PR changes
the default behavior to flashing an empty 4kB image on the other slot,
in order to make sure that the header of any previous image is
overwritten and invalidated.

Change-Id: I26b39f7eada4958f13fe1483fad60c53280d0cdc
Signed-off-by: Jes B. Klinke <jbk@chromium.org>