[rules] Make opentitan_binary top aware
With the current way exec env work, putting all exec env for all
tops in the attributes of the rule is problematic because based
on the current top, certaine exec_env are incompatible (they
indirectly depend on the top's header file, so bazel will complain).

To work around this issue, opentitan_binary is redefined as a macro
that wraps the real rules and transforms the list of exec_env into
a big select() that only picks the execution environments for the
top pointed to by //hw/top. This requires a bit of hackery because
this is a macro, so it's all string matching based on the top
description. See the code for details.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>