[lint] Waive LINE_LENGTH AscentLint warnings in top_darjeeling.sv
top_darjeeling.sv is auto-generated and adhering to the line length
limit is not always feasible for auto-generated files. In particular
when dealing with uniquified and parameterized input/ouput ports and
signal names.

Automatically wrapping those lines would hinder readability and so would
coming up with shorter signal/parameter names (which is especially
challenging given the auto-generation / unification aspects here).

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>