[ci] Add a darjeeling job
This check verifies that the build graph sense when the top is set
to Darjeeling. Furthermore, it tries to build a restricted set of
tests for Darjeeling DV (currently only the uart_smoketest). The
purpose of this check is to prevent regression while the codebase
is being converted to multitop.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>