[testutils] introduce entropy_testutils_entropy_src_init
This initializes entropy src with default configuration.
`entropy_testutils_config_default` is removed.

The idea is that this function will become a no-op for darjeeling when
the user only needs the entropy src to be initialized but doesn't need
special config.

The old `entropy_testutils_config_default`, for most users, are
immediately passed to initialize entropy_src, with only a single user
tweaking it. This user is updated to just explicitly list all config,
and other users are updated to call the new initialization function
instead.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>