[aes,dv] GCM sideload test
This commit allows testing the AES-GCM mode with sideloading.

For this, the try_recover() function was adapted to also handle
GCM messages.

Moreover, a new field output_tag_vld was added - only when this
field is set to 1, the tag comparison is conducted. The reason
for this is that, in sideloading test, the key can get invalid
during a message is processed. Then, the message is split into
two messages. While the second message sees the GCM_TAG item,
the first one does not. Hence, for the first message, do not
conduct the tag comparison.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>