[aes,dv] Make waiting for idle in set_gcm_phase() optional
This commit makes the spinwait for the IDLE status in the
set_gcm_phase() task configurable with the wait_idle bit.

When calling this task in the config_and_transmit() function,
we are already IDLE (as checked by the status_fsm()). To
avoid issues with certain error tests were we actually want
to abort the test, do not use the spinwait in this case.

For other positions in the DV code where we do not use
status_fsm() to check the IDLE state, enable the spinwait
(e.g., in try_recover()).
Also enable waiting for IDLE in the nist test vector gcm
test.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>