[dif,pwrmgr,rstmgr] Add Darjeeling pwrmgr/rstmgr changes
Adds conditional compilation for Darjeeling to match the changed
register definitions, including a change in the number of wakeup reasons
for the rstmgr.

Ideally this can be better modularised through some devicetable
extension API in the future for these blocks, but for now we introduce
`ifdef`s to the DIFs to enable this functionality.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>