[aes,dv] Fix C_DPI input for xcelium
It turns out that xcelium is more picky than vcs when it comes to
uninitialized memory objects. This commit sets the input AAD and
tag to a value if we are not in AES-GCM mode. This is fine as the
C_DPI library anyways ignores those inputs when not in AES-GCM
mode.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>