[aes/dv] Add missing predict() calls to update mirrored RAL values
When using .set() method of the uvm_reg_field base class (this only
updates the desired value in the abstraction class) followed by a
csr_update() call (this updates the DUT in case the desired value
doesn't match the mirrored value), we also need to explicitly update
the mirrored value. This can happen either by using the .predict()
method, or by calling csr_rd() or csr_wr(). This commit adds missing
calls to the .predict() method were required.

Without these calls, the mirrored values can get out of sync and since
in some cases we use the mirrored value to decide whether we need to
update register values, the DUT and the DV environment can get out of
sync.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>