[dif,rstmgr] Add rstmgr DIF for getting SW reset index from DT reset
For space efficiency the rstmgr DT extension only exposes this
information in one direction, mapping from SW reset indexes to DT
resets. This test requires the ability to map from given devices/resets
to the corresponding SW reset index. Implement a DIF function with a
basic linear function for this purpose. Due to the small number of
software resets, this shouldn't introduce a large performance overhead.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>