[gpio] Delay sampling of GPIO inputs to accommodate synchronizers
The inputs were being sampled too soon in `test_gpio_write` on
the expectation that the outputs would all be visible at the
GPIO inputs almost immediately, and that all inputs would change
simultaneously. Even in DV simulation that is not the case because
the CDC modeling introduces single-cycle latency on a random
subset of the input signals.

Introduce a single microsecond delay which, for all practical
clock frequencies, will ensure that the inputs have stabilized by
the time they are sampled.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>