[darjeeling] Move external reset requests into DV environment
Implement the handling of power manager internal request requests
in the DV environment using a dedicated task rather than in the
chiplevel RTL.

This dedicated task may be disabled in the small number of
sequences that wish to supply asynchronous SoC-driven external
reset requests explicitly, such as chip_sw_soc_proxy_smoketest
which now passes.

The original counter logic is maintained in FPGA targets for now
to mitigate the risk of hard-to-diagnose failures if/when Darjeeling
FPGA builds become viable.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>