[uart] Fix uart_tx_rx_alt_clk/rand_baudrate in DV sim
The `sim_dv` targets expect to be able to override the clock
frequency and target baud rate of the UART peripheral under
test in order to cover a wider range of sampling ratios and
phases (= NCO values).

Explicitly widen the parameters to 64 bits before using the
streaming operators because otherwise one simulator replicates
the MS byte whilst another zeroes the additional bytes.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>