[aes/dv] Don't expect to leave GCM_INIT in aes_shadow_reg_errors tests
These tests randomly write the shadow registers inside AES. Until now,
they did not account for the fact that to actually leave the GCM_INIT
phase, the initialization phase needs to complete which is impossible
to hit in these tests. This commit thus aligns the prediction such that
leaving GCM_INIT is not possible in these tests.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>