[reggen] Undetectable zero assignment
While doing FPV analysis on rv_plic, we realised that two variables'
initial assignment to zero could never have an effect. This is because
every bit of the variables gets explicitly written in a continuous
assignment just below.

For example, addr_hit is sized to have indices ${len(regs_flat)}-1:0
and then its bits are assigned in a loop over regs_flat, ensuring all
indices will be assigned.

reg_we_check is sized the same and assigned with an equivalent loop.

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>
Co-authored-by: Rupert Swarbrick <rswarbrick@lowrisc.org>