[site] Fix build-docs for local build
For local build/serve/proxy, the scripts but the book output directly
in the build-site/ directly. But mdbook clears this directly before
outputting any HTML file. As a result, the Doxygen website gets erased.
This commit makes all commands use the same behaviour which is to create
the book in the book/ subdirectory. As a result, for local builds there
is no landing page but the server will still list the root directory
so this shouldn't be an issue.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>