[rv_plic, fpv] Dead Implicit else
de for ip registers is true and since it has software
access as RO, wr_en in prim_subreg is true and it
exist as a condition inside an else_if block which
doesn't have the else block. So, jasper points the
wr_en else_if branch as dead with a description of
implicit branch.

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>