[rv_plic, fpv] Unreachable default case in prim_diff_decode
This default case is unreachable unless we do fault injection. But
with the current standing of prim_diff_decode.sv, it doesn't have
any opinion about what to do next when somebody injects a faulty
state to the FSM. Hence, the FSM would get stuck into the state
unless reset happens.

Even if we add a stopat on state_q, the checker coverage for the
default case will then appear as undetectable asking for an
assertion to mark it checked. But since we get that FSM stuck
into the faulty state, we have no strong opinion about what will
happen if state_q is bad.

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>