[top_earlgrey/doc] Clarify dual core lockstep in block diagram
In the previous version of the block diagram, a single Ibex core was
shown.  This differs from Darjeeling's block diagram as well as from the
interactive block diagrams on https://opentitan.org/documentation, where
the second core of the dual lockstep configuration was shown.

This commit adds the second core to the block diagram in Earl Grey's
datasheet, to clarify that Earl Grey operates Ibex in dual core lockstep
mode.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>