[prim,dv] Fix problem with forcing in prim_sparse_fsm_flop_if
The code in this interface was a bit confused: the restore_fault()
task would only do anything if CustomForceName was empty, but then it
did something that was guaranteed not to work (it was depositing to an
invalid HDL path).

Presumably, that means this interface has only ever been used with a
nonempty CustomForceName. Delete the code that's obviously wrong (and
presumably untested).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>