[base] Increase the period of decoy pairs in `hardened_memory.c`.
Adds a word-offset of 1 to the index of the second decoy in `hardened_memcpy`
so that we have more possible source/destination pairs. Otherwise, there would
only be 4 possible unique pairs; 4 word-sized increments after `decoy[3] =
decoy[7]` would be `decoy[7] = decoy[3]`. Adding an offset results in more
unique pairs for minimal overhead.

This doesn't really have an effect on `hardened_memeq` since all the decoys
have to be equal there, but I changed it there anyway just to keep the pattern
consistent.

Signed-off-by: Jade Philipoom <jadep@zerorisc.com>