[kmac,dv] Shorten messages for stress all test
Some tests are failing because they run into a timeout. This
timeout is caused by a process that is waiting for the tb to
stop writing to the msg fifo. Since the messages can be very long
the wait runs into a timeout and the test fails. This commit fixes
this issue.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>