[edn/doc] Document how to uninstantiate a CSRNG instance through EDN
If firmware just disables EDN, there is a high risk of leaving the
associated CSRNG instance in an undetermined state. Getting out of this
state means disabling and re-enabling and thus reconfiguring both EDNs
and all CSRNG instances which is non-ideal.

This document, how firmware can safely uninstantiate a CSRNG instance
through EDN without to enable reconfiguring individual EDNs.

This is related to lowRISC/OpenTitan#26781.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>