[dpi] Avoid undefined behaviour in uartdpi.c
It turns out the cfmakeraw function is supposed to be used to
customise an initialised termios struct, rather than writing the whole
thing. Do this by grabbing values from the configuration of the slave
side of the uart pty.

(The name is confusing! It should be read as "make the thing more raw"
rather than "make a raw thing")

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>