[rom_ctrl,rtl] Add a flop between rom_ctrl and kmac
This injects a clock edge between data coming out of the ROM and data
going into KMAC to be hashed. Inserting the delay is pretty trivial:
we just use a prim_fifo_sync.

There *is* an area vs. efficiency question. Because there is no
pass-through, a prim_fifo_sync with depth 1 will have half throughput,
because it has to alternate between taking an item and passing it on.
Although, there's the bandwidth to do both on the same cycle, breaking
the combinatorial path means that we can't pass the KMAC block's
"ready" signal all the way back to the ROM.

This works perfectly well, but takes twice as long as you might hope.

A lazy workaround is to use Depth = 2. This is a little silly, because
it uses twice as much area as necessary. But it's very easy from a
coding perspective!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>