[i2c, test_utils] Provide SDA rise/fall when setting I2C speed
Currently, the SDA rise and fall time is hardcoded to 400/110
nanoseconds respectively, despite the fact that these two values are
dependent on physical characteristics. Specifically the signal rise time
is dependent upon the capacitance and physical design of the I2C bus,
and the fall time is a function of the SCL driver which manages the slew
rate through the falling edge.

In reality, these values can likely be much lower and may differ between
execution environments (kDeviceType). This commit refactors the
`i2c_testutils_set_speed` function to allow a test to provide the
rise/fall times that it wishes to use (as the test knows what is
connected), and just uses the default values in all cases for now.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>