[entropy_src/rtl] Fix bucket HT result for NumBucketHtInst > 1
Before this commit, the max bin counter values of the individual
NumBucketHtInst groups of BucketHtDataWidth channels got added together.
This is not meaningful as it attenuates outliers observerd in individual
groups. Instead, only the max value of the the NumBucketHtInst groups
should be returned. This value is most meaningful for defining the
thresholds provided by firmware.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>