[adc_ctrl,dv] Expand comments a bit and tidy up adc_ctrl_fsm_sva_if
There is no functional change, but dropping the ASSERT macro makes
things a little easier to read.

The hardware reset assertions should now be a little more
efficient. I don't think we really need to bother asserting that the
signals don't magically start changing somewhere in the middle of a
reset(!), so I've restructured them slightly to check the first clock
after the start of a reset.

The first syntax I used for this (which is equivalent to the previous
version, but without noise from the macro) uses two "clocks" and is of
the form

  @(negedge rst_aon_ni) 1 |=> @(posedge clk_aon_i) XYZ

Here, the |=> waits for the event on the right hand side, so it means
"On the first clock edge after reset is asserted, XYZ". Very nice!

Unfortunately, this isn't actually supported by Verible's
SystemVerilog front-end. Grr! I was very annoyed at first, but
realised that there's actually an even cleaner way to do it. Because
this is a bound-in interface, we can easily define a helper signal.

This commit defines a "start_of_reset" flag, which set when reset is
asserted and then cleared on the first clock. Now, the properties can
be very simple, looking like:

  start_of_reset -> XYZ

(checked on the positive edge of clk_aon_i because of the default
clocking statement).

That's even shorter! Yippee!

When I initially wrote this, I worried about the scheduling because we
clear start_of_reset on the clock edge and look at the value at the
same time. Fortunately, this is ok: the value tested in the expression
is its value from the preponed region (see IEEE 1800, 16.5.1), so
we'll get the value that hasn't yet been cleared.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>