[adc_ctrl,dv] Minor tweak to adc_ctrl_fsm_reset_vseq
At the start of this sequence it tries to set the filters to always
match. It does so by making the range [0, 1023] (the whole input
space) but then sets the cond flag to 1. That means that they will
match values that are not in [0, 1023]...

Invert the meaning. Rather oddly, the test seems to have worked
equally well either way, so maybe this was completely pointless! But
at least the comments and code are now consistent.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>